ONE-27 change to store episode on django admin page
