--- templates/admin/store.html
@@ -1,26 +1,0 @@
-{% extends 'base.html' %}
-
-
-{% block body %}
-<h1>Admin</h1>
-<form action="{% url 'admin' %}" method="POST" enctype='multipart/form-data'>
-  {% csrf_token %}
-  <p>
-    <div><label for="">ラジオID</label></div>
-    <input type="number" name="id">
-  </p>
-  <p>
-    <div><label for="">回</label></div>
-    <input type="number" name="number">
-  </p>
-  <p>
-    <div><label for="">オーディオファイル</label></div>
-    <input type="file" name="audio_file">
-  </p>
-  <p>
-    <div><label for="">放送日</label></div>
-    <input type="date" name="air_date">
-  </p>
-  <p><input type="submit" value="登録"></p>
-</form>
-{% endblock body %}
